// source/frontend/src/models/api/generated/ApiGen_Concepts_ExpropriationEvent.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDate } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Base_CodeType } from './ApiGen_Base_CodeType';
import { ApiGen_Concepts_AcquisitionFileOwner } from './ApiGen_Concepts_AcquisitionFileOwner';
import { ApiGen_Concepts_InterestHolder } from './ApiGen_Concepts_InterestHolder';

// LINK: @backend/apimodels/Models/Concepts/ExpropriationEvent/ExpropriationEventModel.cs
export interface ApiGen_Concepts_ExpropriationEvent extends ApiGen_Base_BaseAudit {
  id: number | null;
  acquisitionFileId: number;
  acquisitionOwnerId: number | null;
  acquisitionOwner: ApiGen_Concepts_AcquisitionFileOwner | null;
  interestHolderId: number | null;
  interestHolder: ApiGen_Concepts_InterestHolder | null;
  eventType: ApiGen_Base_CodeType<string> | null;
  eventDate: UtcIsoDate | null;
}